feat: cycle explorer layout with Shift+F instead of dedicated key
Fold the hide-left-sidebar toggle into Shift+F as a 3-phase layout
cycle for the three-pane explorer: normal -> sidebar hidden ->
fullscreen -> normal. Drop the dedicated kb.HideSidebar (`|`) binding.
Dashboards keep their plain fullscreen toggle. Status messages share a
consistent "Layout: ..." prefix.

Keep both app.go and tabs.go under the 800-line revive cap (breached
after merging main) by extracting cohesive groups into new files,
mirroring the existing securityModelState pattern:
- easter_egg_state.go: embedded easterEggState sub-struct
- tabs_portforward.go: port-forward list helpers

Update tests, README, docs/keybindings.md, docs/config-reference.md,
docs/config.schema.json, and the in-app help.
